requirements.txt
# Schwab API Streaming Dependencies
schwab-py>=1.0.0
python-dotenv>=1.0.0
asyncio
websockets>=12.0
httpx>=0.25.0
pydantic>=2.0.0
